macro traits

+++ graph/two_sat.h
@@ -19,41 +19,40 @@
 struct two_sat {
   graph_t<void> graph;
   std::vector<char> truth;  // no vector<bool>
-  two_sat(int n): graph(2*n), truth(n) {}
+  two_sat(int n) : graph(2 * n), truth(n) {}
   void reset(int n) {
-    graph.reset(2*n);
+    graph.reset(2 * n);
     truth.resize(n);
   }
   bool operator[](int i) const { return truth[i]; }
   bool solve() {
     strongly_connected scc(graph);
-    for (int i = 0; i < truth.size(); i++) {
-      if (scc[2*i] == scc[2*i + 1]) return false;
-      truth[i] = (scc[2*i] < scc[2*i + 1]);
+    for (int i = 0; i < (int)truth.size(); i++) {
+      if (scc[2 * i] == scc[2 * i + 1]) return false;
+      truth[i] = (scc[2 * i] < scc[2 * i + 1]);
     }
     return true;
   }
 
   void either(int a, int b) {
-    graph.add_arc(a^1, b);
-    graph.add_arc(b^1, a);
+    graph.add_arc(a ^ 1, b);
+    graph.add_arc(b ^ 1, a);
   }
   void implies(int a, int b) {
     graph.add_arc(a, b);
-    graph.add_arc(b^1, a^1);
+    graph.add_arc(b ^ 1, a ^ 1);
   }
   void iff(int a, int b) {
     implies(a, b);
     implies(b, a);
   }
   void either(int a, bool a_truth, int b, bool b_truth) {
-    either(2*a + !a_truth, 2*b + !b_truth);
+    either(2 * a + !a_truth, 2 * b + !b_truth);
   }
   void implies(int a, bool a_truth, int b, bool b_truth) {
-    implies(2*a + !a_truth, 2*b + !b_truth);
+    implies(2 * a + !a_truth, 2 * b + !b_truth);
   }
   void iff(int a, bool a_truth, int b, bool b_truth) {
-    iff(2*a + !a_truth, 2*b + !b_truth);
+    iff(2 * a + !a_truth, 2 * b + !b_truth);
   }
 };
-
